move the stream and stream-browser to somewhere better
