+++ lib/internal/streams/stream-browser.js
