+++ lib/internal/streams/stream.js
