+++ package.json
@@ -43,7 +43,7 @@
     "./readable.js": "./readable-browser.js",
     "./writable.js": "./writable-browser.js",
     "./duplex.js": "./duplex-browser.js",
-    "./_internal_stream.js": "./_internal_stream-browser.js"
+    "./lib/internal/streams/stream.js": "./lib/internal/streams/stream-browser.js"
   },
   "react-native": {
     "stream": false
